OpenAM changes, and bridging gap between zfcuser

+++ Common/src/Common/Controller/LogoutController.php
@@ -5,6 +5,7 @@
  */
 namespace Common\Controller;
 
+use Dvsa\Olcs\Utils\Auth\AuthHelper;
 use Zend\Mvc\Controller\AbstractActionController;
 use Zend\Session\Container;
 
@@ -15,6 +16,10 @@ class LogoutController extends AbstractActionController
 {
     public function indexAction()
     {
+        if (AuthHelper::isOpenAm() === false) {
+            return $this->redirect()->toRoute('zfcuser/logout');
+        }
+
         $session = new Container();
         $session->getManager()->destroy(['clear_storage' => true, 'send_expire_cookie' => true]);
